% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_single_tree.R
\name{build_single_tree}
\alias{build_single_tree}
\title{Build single tree of a forest}
\usage{
build_single_tree(X, Y, mtree, min_leaf, V_inv, Command)
}
\arguments{
\item{X}{Input Training matrix of M x N, M is the number of training samples and N is the number of features}

\item{Y}{Output Training response of M x T, M is the number of samples and T is number of ouput Features(Response)}

\item{mtree}{number of features will be used for each split}

\item{min_leaf}{minimum number of samples in the leaf node}

\item{V_inv}{Covariance matrix of response matrix}

\item{Command}{1 for RF and 2 for MRF}
}
\value{
Model of single tree of the forest
}
\description{
Single tree using the training samples has been built and has been returned as a Model
}

