% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Imputation.R
\name{Imputation}
\alias{Imputation}
\title{Imputation of a vector of number}
\usage{
Imputation(XX)
}
\arguments{
\item{XX}{a vector of N x 1}
}
\value{
Imputed vector of N x 1
}
\description{
Imputes the values of the vector which are NaN
}
\details{
If a value is missing, it will be replaced by an imputed value which is an average of previous and next value. While if previous 
or next value is also missing, the closet value has been used as an imputed value.
}

