% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predicting.R
\name{predicting}
\alias{predicting}
\title{Prediction of testing sample in a node}
\usage{
predicting(Single_Model, i, X_test, Variable_number)
}
\arguments{
\item{Single_Model}{Model of a particular tree}

\item{i}{Number of split. Used as an index, which indicates where in the list the splitting
criteria of this split has been stored.}

\item{X_test}{Testing samples of size Q x N, Q is the number of testing samples and N is the number of features (same order and
size used as training)}

\item{Variable_number}{Number of Output Features}
}
\value{
Prediction result of a testing samples in a node
}
\description{
Provides the value of a testing sample in a node which refers to which child node it will go using the splitting criteria 
of the tree node or prediction results if the node is a leaf.
}
\details{
The function considers the output at a particular node. If the node is a leaf, the average of output responses 
is returned as prediction result. For a non-leaf node, the direction of left or right node is decided based on 
the node threshold and splitting feature value.
}
