% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/single_tree_prediction.R
\name{single_tree_prediction}
\alias{single_tree_prediction}
\title{single_tree_prediction}
\usage{
single_tree_prediction(Single_Model, X_test, Variable_number)
}
\arguments{
\item{Single_Model}{Model of a single particular tree}

\item{X_test}{Testing samples of Q x N, Q is the number of testing samples and N is the number of features(same order and
size used as training)}

\item{Variable_number}{Number of drugs which response needs to be calculated}
}
\value{
predicted response of Testing samples for a single tree
}
\description{
Using the model of the tree, for testing samples, prediction has been calculated
}
\details{
A model contrains for each split of the node what criteria has been used. For the testing samples, using the criteria
testing sample reaches a leaf node, which have some response value stored. Average of these values are used as prediction
for the testing samples
}

