% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spliting.R
\name{spliting}
\alias{spliting}
\title{spliting}
\usage{
spliting(X, Y, mtree, Index, V_inv, Command)
}
\arguments{
\item{X}{Input Training matrix of M x N, M is the number of training samples and N is the number of features}

\item{Y}{Output Training response of M x T, M is the number of samples and T is number of ouput Features(Response)}

\item{mtree}{number of features will be used for each split}

\item{Index}{Index of training samples}

\item{V_inv}{Covariance matrix of response matrix}

\item{Command}{1 for RF and 2 for MRF}
}
\value{
Spliting criteria for the splitting of the node
}
\description{
Spliting of training samples in a node using different feature vector has been calculated
}
\details{
In a node of training samples, which spliting gives best cost reduction has been calculted. mtree number of feature
are considered and for a specific feature and specific split point, leaf node and right node split cost has been calculated.
Whichever feature and split has been given minimum cost has been selected and the criteria are returned
}

