% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_single_tree.R
\name{build_single_tree}
\alias{build_single_tree}
\title{Model of Single tree of Random Forest or Multivariate Random Forest}
\usage{
build_single_tree(X, Y, mtree, min_leaf, V_inv, Command)
}
\arguments{
\item{X}{Input Training matrix of M x N, M is the number of training samples and N is the number of features}

\item{Y}{Output Training response of M x T, M is the number of samples and T is number of ouput Features(Response)}

\item{mtree}{Number of randomly selected features used for each split}

\item{min_leaf}{Minimum number of samples in the leaf node}

\item{V_inv}{Covariance matrix of Output Feature matrix}

\item{Command}{1 for RF and 2 for MRF depending on the method}
}
\value{
Model of single tree of the forest(RF or MRF)
}
\description{
Build the model of a tree of RF or MRF using the training samples, which is used for the prediction of testing samples
}
\details{
The Model contains the list of lists, where each list contains either the splitting criteria of a split in
a node or, output feature matrix of the samples of a node depending upon which node the list is representing.
}

