% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CrossValidation.R
\name{CrossValidation}
\alias{CrossValidation}
\title{cross validation}
\usage{
CrossValidation(X, Y, F)
}
\arguments{
\item{X}{M x N Input matrix, M is the number of samples and N is the number of features}

\item{Y}{output Response as column vector}

\item{F}{Number of Fold in cross validation}
}
\value{
list with the following components: 
\item{TrainingData}{List of matices with matrix containing Cross Validates Training Data, where number of list
equal to user defined cross validation}
\item{TestingData}{List of matices with matrix containing Cross Validates Testing Data, where number of list
equal to user defined cross validation}
\item{OutputTrain}{List of matices with matrix containing Cross Validates Training Response Data, where number of list
equal to user defined cross validation}
\item{OutputTest}{List of matices with matrix containing Cross Validates Testing Response Data, where number of list
equal to user defined cross validation}
\item{FoldedIndex}{Index of Different Fold}
}
\description{
Generates Cross validated Matices and Output Response, where number of fold in cross validation is user defined
}

