% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Imputation.R
\name{Imputation}
\alias{Imputation}
\title{Imputation of a numerical vector}
\usage{
Imputation(XX)
}
\arguments{
\item{XX}{a vector of size N x 1}
}
\value{
Imputed vector of size N x 1
}
\description{
Imputes the values of the vector that are NaN
}
\details{
If a value is missing, it will be replaced by an imputed value that is an average of previous and 
next value. If previous or next value is also missing, the closest value is used as the imputed value.
}
