% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Imputation.R
\name{Imputation}
\alias{Imputation}
\title{Imputation of a output response}
\usage{
Imputation(XX)
}
\arguments{
\item{XX}{Vector of response which needs to be imputed}
}
\value{
Imputed values of that vector
}
\description{
In case some responses are missing, this function will impute those values.
}
\details{
If a value is missing, it will be replaced by an imputed value which is an average of previous and next value. While if previous 
or next value is also missing, the closet value has been used as an imputed value.
}

