% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Multi_D_mod.R
\name{Multi_D_mod}
\alias{Multi_D_mod}
\title{Information Gain or splitting cost}
\usage{
Multi_D_mod(y, V_inv, Command)
}
\arguments{
\item{y}{Feature Vector}

\item{V_inv}{Covariance matrix of response matrix}

\item{Command}{1 for RF and 2 for MRF}
}
\value{
threshold_value
}
\description{
For a specific spliting in a node for a spcific feature vector, cost has been calculated.
}

