% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predicting.R
\name{predicting}
\alias{predicting}
\title{Prediction of testing sample}
\usage{
predicting(Single_Model, i, X_test, Variable_number)
}
\arguments{
\item{Single_Model}{Model of a single particular tree}

\item{i}{number of which split needs to be done}

\item{X_test}{Testing samples of Q x N, Q is the number of testing samples and N is the number of features(same order and
size used as training)}

\item{Variable_number}{How many drugs are given for computation}
}
\value{
Predicted response of a Testing samples
}
\description{
For a specific node, which way should the testing sample goes has been determined using spliting criteria stored in the model.
}

