% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/single_tree_prediction.R
\name{single_tree_prediction}
\alias{single_tree_prediction}
\title{Prediction of Testing Samples for single tree}
\usage{
single_tree_prediction(Single_Model, X_test, Variable_number)
}
\arguments{
\item{Single_Model}{Random Forest or Multivariate Random Forest Model of a particular tree}

\item{X_test}{Testing samples of Q x N, Q is the number of testing samples and N is the number of features(same order and
size used as training)}

\item{Variable_number}{Number of Output Features}
}
\value{
Prediction result of the Testing samples for a particular tree
}
\description{
Predicts the output features of testing samples using the model of a tree
}
\details{
A model contrains splitting criteria for all the split of the tree and output features of training samples in the leaf nodes.
A testing sample using these criteria will go to a leaf node and average of the output feature vectors in the leaf node
is considered as the prediction of that testing sample.
}

