% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_node.R
\name{split_node}
\alias{split_node}
\title{split_node}
\usage{
split_node(X, Y, mtree, Index, i, model, min_leaf, V_inv, Command)
}
\arguments{
\item{X}{Input Training matrix of M x N, M is the number of training samples and N is the number of features}

\item{Y}{Output Training response of M x T, M is the number of samples and T is number of ouput Features(Response)}

\item{mtree}{number of features will be used for each split}

\item{Index}{Index of training samples}

\item{i}{number of split}

\item{model}{The list of numbers where different spliting criteria has been stored}

\item{min_leaf}{minimum number of samples in the leaf node}

\item{V_inv}{Covariance matrix of response matrix}

\item{Command}{1 for RF and 2 for MRF}
}
\value{
Model Updated Model after split has been done
}
\description{
The Splitting criteria in a node of a tree has been calculated
}

